missions/mission_17/requirements.txt
pandas
numpy
pillow
Pillow
timm

streamlit
#streamlit-drawable-canvas

torch
transformers
huggingface_hub[hf_xet]
qwen-vl-utils
accelerate